fix(org-sdk): identify a demand by a minted id, not by its address
`DemandState` and `Outcome::Certified` identified the demand they describe with
`Arc::as_ptr(demand) as usize`, while holding no `Arc` to keep it alive. That is
an ABA hazard: once the demand is dropped a new one can be allocated exactly
where it lived, and with the same capability and population it compares equal to
the record describing its predecessor - so a needed convergence is skipped, or,
symmetrically, a degraded state is wrongly paced.

Core now mints a process-wide monotone id per published demand and exposes it as
`OrgSensingCapabilityDemand::id`. One `u64`, and it cannot alias whatever the
allocator does. The witness observable `sensing_demand_state` reads it too - it
exists precisely to tell "the same demand was reused" from "an identical one was
re-acquired", which the address could not answer honestly.

The witness churns retire/re-retain in a loop, which is the shape that frees an
allocation for immediate reuse, and asserts the identities are both unique and
monotone.

+++ net/crates/net/sdk/tests/org_exact_sensing.rs
@@ -399,7 +399,7 @@ fn release_sensing_capacity(node: &Arc<MeshNode>, tickets: Vec<SensingLeaseTicke
 }
 
 /// This capability's retained population, holders and demand identity.
-fn demand_state(client: &OrgClient) -> Option<(Vec<u64>, Vec<u64>, usize)> {
+fn demand_state(client: &OrgClient) -> Option<(Vec<u64>, Vec<u64>, u64)> {
     client
         .sensing_demand_state(&capability())
         .map(|(mut population, mut retained, identity)| {
